include/interface/agent.h: Move to "include/laik/agent.h".
